pyproject.toml
[tool.poetry]
name = "data-engineer-jobs-scraping"
version = "0.1.0"
description = ""
authors = ["Guilherme Samuel <guilhermesamuel79@gmail.com>"]
readme = "README.md"
packages = [{include = "data_engineer_jobs_scraping", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.27.1"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
sqlalchemy = "^2.0.36"
duckdb = "^1.1.3"
duckdb-engine = "^0.14.0"
numpy = "^2.2.1"
pandas = "^2.2.3"
elasticsearch = "^8.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
